Add some pagination , fix some issues and hope it still works

+++ app/Http/Traits/ImagesTrait.php
@@ -9,7 +9,7 @@ trait ImagesTrait{
     {
         $file->move(public_path('uploads/'.$path), $fileNewName);
         $NewFilePath = 'uploads/'.$path. '/'.$fileNewName;
-        if(!is_null($oldFilePath))
+        if(!is_null($oldFilePath) && file_exists($oldFilePath))
         {
             unlink(public_path($oldFilePath));
         }
